+++ GP4-Challenge-2/Views/CategoriesViews/SliderCategoriesView.swift
